Added logs
